Fix volumes variable
In Docker, the /docker-entrypoint-initdb.d directory is a location where you can place SQL scripts or other initialization scripts that you want to be executed when a Docker container is started for the first time. These scripts are typically used to initialize a database with schema, tables, and initial data.